src/main/webapp/app/entities/don/route/don-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IDon, Don } from '../don.model';
import { DonService } from '../service/don.service';

import { DonRoutingResolveService } from './don-routing-resolve.service';

describe('Service Tests', () => {
  describe('Don routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: DonRoutingResolveService;
    let service: DonService;
    let resultDon: IDon | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(DonRoutingResolveService);
      service = TestBed.inject(DonService);
      resultDon = undefined;
    });

    describe('resolve', () => {
      it('should return IDon returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultDon = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultDon).toEqual({ id: 123 });
      });

      it('should return new IDon if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultDon = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultDon).toEqual(new Don());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Don })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultDon = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultDon).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
